move json diff code into library
